+++ src/test/java/org/folio/edge/sip2/api/PatronStatusIT.java
@@ -28,6 +28,7 @@ class PatronStatusIT extends AbstractErrorDetectionEnabledTest {
       "/wiremock/stubs/mod-users-bl/200-get-user-by-id.json",
       "/wiremock/stubs/mod-fee-fines/200-get-accounts-open-status.json",
       "/wiremock/stubs/mod-fee-fines/200-get-manualblocks.json",
+      "/wiremock/stubs/mod-fee-fines/200-get-automated-patron-blocks.json",
   })
   void getPatronStatus_noManualBlocks_patronStatusFlagsAreEmpty() throws Throwable {
     executeInSession(
@@ -57,6 +58,7 @@ class PatronStatusIT extends AbstractErrorDetectionEnabledTest {
       "/wiremock/stubs/mod-users-bl/200-get-user-by-id.json",
       "/wiremock/stubs/mod-fee-fines/200-get-accounts-open-status.json",
       "/wiremock/stubs/mod-fee-fines/200-get-manualblocks-with-borrowing-block.json",
+      "/wiremock/stubs/mod-fee-fines/200-get-automated-patron-blocks.json",
   })
   void getPatronStatus_withManualBorrowingBlock_allStatusFlagsSetAndMessageReturned()
       throws Throwable {
@@ -77,6 +79,39 @@ class PatronStatusIT extends AbstractErrorDetectionEnabledTest {
               assertThat(response.getScreenMessage())
                   .isEqualTo(List.of(
                       "Your account is blocked. Please contact the library."));
+            }
+        ));
+  }
+
+  @Test
+  @WiremockStubs({
+      "/wiremock/stubs/mod-settings/200-get-locale.json",
+      "/wiremock/stubs/mod-settings/200-get-settings.json",
+      "/wiremock/stubs/mod-login/201-post-acs-login.json",
+      "/wiremock/stubs/mod-users/200-get-user-by-patron-identifier.json",
+      "/wiremock/stubs/mod-users-bl/200-get-user-by-id.json",
+      "/wiremock/stubs/mod-fee-fines/200-get-accounts-open-status.json",
+      "/wiremock/stubs/mod-fee-fines/200-get-manualblocks.json",
+      "/wiremock/stubs/mod-fee-fines/200-get-automated-patron-blocks-with-borrowing-block.json",
+  })
+  void getPatronStatus_withAutomatedBorrowingBlock_allStatusFlagsSetAndMessageReturned()
+      throws Throwable {
+    executeInSession(
+        successLoginExchange(),
+        sip2Exchange(
+            Sip2Commands.patronStatus(PATRON_BARCODE),
+            sip2Result -> {
+              assertSuccessfulExchange(sip2Result);
+
+              var respMsg = sip2Result.getResponseMessage();
+              assertThat(respMsg).startsWith("24");
+
+              var response = parseResponse(respMsg);
+              assertThat(response.getValidPatron()).isTrue();
+              assertThat(response.getPatronStatus())
+                  .isEqualTo(EnumSet.allOf(PatronStatus.class));
+              assertThat(response.getScreenMessage())
+                  .isEqualTo(List.of("Patron has too many items checked out"));
             }
         ));
   }