+++ NEWS.md
@@ -5,6 +5,7 @@
 * Use GitHub Workflows for Maven ([SIP2-307](https://folio-org.atlassian.net/browse/SIP2-307))
 * Fix misleading log messages when token refresh fails and recovery login is used([SIP2-316](https://folio-org.atlassian.net/browse/SIP2-316))
 * Fix locale agnostic number formatting ([SIP2-300](https://folio-org.atlassian.net/browse/SIP2-300))
+* Add automated patron blocks support to patron status/information response ([SIP2-313](https://folio-org.atlassian.net/browse/SIP2-313))
 
 ---